Move content from project/models to new submission.py and coreproject.py files
