+++ physionet-django/project/models.py
@@ -3,7 +3,6 @@ import hashlib
 from html import unescape
 import os
 import shutil
-import uuid
 import pytz
 import logging
 from distutils.version import StrictVersion
@@ -25,7 +24,8 @@ from project.modelcomponents.authors import *
 from project.modelcomponents.access import *
 from project.modelcomponents.fields import *
 from project.modelcomponents.generic import *
-from project.quota import DemoQuotaManager
+from project.modelcomponents.submission import *
+from project.modelcomponents.coreproject import *
 from project.utility import (get_tree_size, get_file_info, get_directory_info,
                              list_items, StorageInfo,
                              clear_directory, LinkFilter)
@@ -189,58 +189,6 @@ class PublishedPublication(BasePublication):
     """
     project = models.ForeignKey('project.PublishedProject',
         db_index=True, related_name='publications', on_delete=models.CASCADE)
-
-
-class CoreProject(models.Model):
-    """
-    The core underlying object that links all versions of the project in
-    its various states
-    """
-    id = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
-    creation_datetime = models.DateTimeField(auto_now_add=True)
-    # doi pointing to the latest version of the published project
-    doi = models.CharField(max_length=50, blank=True, null=True)
-    # Maximum allowed storage capacity in bytes.
-    # Default = 100Mb. Max = 10Tb
-    storage_allowance = models.BigIntegerField(default=104857600,
-        validators=[MaxValueValidator(109951162777600),
-                    MinValueValidator(104857600)])
-
-    def active_new_version(self):
-        "Whether there is a new version being worked on"
-        return bool(self.activeprojects.filter())
-
-    def get_published_versions(self):
-        """
-        Return a queryset of PublishedProjects, sorted by version.
-        """
-        return self.publishedprojects.filter().order_by('version_order')
-
-    @property
-    def total_published_size(self):
-        """
-        Total storage size of the published projects.
-
-        This is the sum of the PublishedProjects'
-        incremental_storage_size values (which will be less than the
-        sum of the main_storage_size values if there are files that
-        are shared between versions) and reflects the actual size of
-        the data on disk.
-        """
-        result = self.publishedprojects \
-                     .aggregate(total=models.Sum('incremental_storage_size'))
-        # The sum will be None if there are no publishedprojects.  It will
-        # also be None if any projects have incremental_storage_size=None.
-        return result['total'] or 0
-
-
-class ProjectType(models.Model):
-    """
-    The project types available on the platform
-    """
-    id = models.PositiveSmallIntegerField(primary_key=True)
-    name = models.CharField(max_length=20)
-    description = models.TextField()
 
 
 class Metadata(models.Model):
@@ -811,81 +759,6 @@ class Metadata(models.Model):
             citation_dict[style] = self.citation_text(style)
 
         return citation_dict
-
-
-class SubmissionInfo(models.Model):
-    """
-    Submission information, inherited by all projects.
-
-    Every project (ActiveProject, PublishedProject, and
-    ArchivedProject) inherits from this class as well as Metadata.
-    The difference is that the fields of this class contain internal
-    information about the publication process; Metadata contains the
-    public information that will be shown on the published project
-    pages.
-
-    In particular, UnpublishedProject.modified_datetime will be
-    updated when any field of Metadata is altered (see
-    UnpublishedProject.save), but not when a field of SubmissionInfo
-    is modified.
-
-    New fields should be added to this class only if they do not
-    affect the content of the project as it will be shown when
-    published.
-    """
-
-    editor = models.ForeignKey('user.User',
-        related_name='editing_%(class)ss', null=True,
-        on_delete=models.SET_NULL, blank=True)
-    # The very first submission
-    submission_datetime = models.DateTimeField(null=True, blank=True)
-    author_comments = models.CharField(max_length=20000, default='', blank=True)
-    editor_assignment_datetime = models.DateTimeField(null=True, blank=True)
-    # The last revision request (if any)
-    revision_request_datetime = models.DateTimeField(null=True, blank=True)
-    # The last resubmission (if any)
-    resubmission_datetime = models.DateTimeField(null=True, blank=True)
-    editor_accept_datetime = models.DateTimeField(null=True, blank=True)
-    # The last copyedit (if any)
-    copyedit_completion_datetime = models.DateTimeField(null=True, blank=True)
-    author_approval_datetime = models.DateTimeField(null=True, blank=True)
-
-    # When the submitting project was created
-    creation_datetime = models.DateTimeField(auto_now_add=True)
-
-    edit_logs = GenericRelation('project.EditLog')
-    copyedit_logs = GenericRelation('project.CopyeditLog')
-
-    # For ordering projects with multiple versions
-    version_order = models.PositiveSmallIntegerField(default=0)
-
-    # Anonymous access
-    anonymous = GenericRelation('project.AnonymousAccess')
-
-    class Meta:
-        abstract = True
-
-    def quota_manager(self):
-        """
-        Return a QuotaManager for this project.
-
-        This can be used to calculate the project's disk usage
-        (represented by the bytes_used and inodes_used properties of
-        the QuotaManager object.)
-        """
-        allowance = self.core_project.storage_allowance
-        published = self.core_project.total_published_size
-        limit = allowance - published
-
-        # DemoQuotaManager needs to know the project's toplevel
-        # directory as well as its creation time (so that files
-        # present in multiple versions can be correctly attributed to
-        # the version where they first appeared.)
-        quota_manager = DemoQuotaManager(
-            project_path=self.file_root(),
-            creation_time=self.creation_datetime)
-        quota_manager.set_limits(bytes_hard=limit, bytes_soft=limit)
-        return quota_manager
 
 
 class UnpublishedProject(models.Model):
@@ -1939,17 +1812,6 @@ def exists_project_slug(slug):
             or PublishedProject.objects.filter(slug=slug))
 
 
-class ProgrammingLanguage(models.Model):
-    """
-    Language to tag all projects
-    """
-    name = models.CharField(max_length=50, unique=True)
-
-    def __str__(self):
-        return self.name
-
-
-
 class StorageRequest(BaseInvitation):
     """
     A request for storage capacity for a project
@@ -1964,139 +1826,6 @@ class StorageRequest(BaseInvitation):
     def __str__(self):
         return '{0}GB for project: {1}'.format(self.request_allowance,
                                                self.project.__str__())
-
-
-class EditLog(models.Model):
-    """
-    Log for an editor decision. Also saves submission info.
-    """
-    # Quality assurance fields for data and software
-    QUALITY_ASSURANCE_FIELDS = (
-        # 0: Database
-        ('soundly_produced', 'well_described', 'open_format',
-         'data_machine_readable', 'reusable', 'no_phi', 'pn_suitable'),
-        # 1: Software
-        ('soundly_produced', 'well_described', 'open_format', 'no_phi',
-            'reusable', 'pn_suitable'),
-        # 2: Challenge
-        ('soundly_produced', 'well_described', 'open_format',
-         'data_machine_readable', 'reusable', 'no_phi', 'pn_suitable'),
-        # 3: Model
-        ('soundly_produced', 'well_described', 'open_format',
-         'data_machine_readable', 'reusable', 'no_phi', 'pn_suitable'),
-    )
-    # The editor's free input fields
-    EDITOR_FIELDS = ('editor_comments', 'decision', 'auto_doi')
-
-    COMMON_LABELS = {
-        'reusable': 'Does the project include everything needed for reuse by the community?',
-        'pn_suitable': 'Is the content suitable for PhysioNet?',
-        'editor_comments': 'Comments to authors',
-        'no_phi': 'Is the project free of protected health information?',
-        'data_machine_readable': 'Are all files machine-readable?'
-    }
-
-    LABELS = (
-        # 0: Database
-        {'soundly_produced': 'Has the data been produced in a sound manner?',
-         'well_described': 'Is the data adequately described?',
-         'open_format': 'Is the data provided in an open format?',
-         'data_machine_readable': 'Are the data files machine-readable?'},
-        # 1: Software
-        {'soundly_produced': 'Does the software follow best practice in development?',
-         'well_described': 'Is the software adequately described?',
-         'open_format': 'Is the software provided in an open format?'},
-        # 2: Challenge
-        {'soundly_produced': 'Has the challenge been produced in a sound manner?',
-         'well_described': 'Is the challenge adequately described?',
-         'open_format': 'Is all content provided in an open format?'},
-        # 3: Model
-        {'soundly_produced': 'Does the software follow best practice in development?',
-         'well_described': 'Is the software adequately described?',
-         'open_format': 'Is the software provided in an open format?'},
-    )
-
-    HINTS = {
-        'no_phi': [
-            'No dates in WFDB header files (or anonymized dates only)?',
-            'No identifying information of any individual'
-            ' (caregivers as well as patients)?',
-            'No ages of individuals above 89 years?',
-            'No hidden metadata (e.g. EDF headers)?',
-            'No internal timestamps, date-based UUIDs or other identifiers?',
-        ],
-        'open_format': [
-            'No compiled binaries or bytecode?',
-            'No minified or obfuscated source code?',
-        ],
-    }
-
-    content_type = models.ForeignKey(ContentType, on_delete=models.CASCADE)
-    object_id = models.PositiveIntegerField()
-    project = GenericForeignKey('content_type', 'object_id')
-
-    # When the submitting author submits/resubmits
-    submission_datetime = models.DateTimeField(auto_now_add=True)
-    is_resubmission = models.BooleanField(default=False)
-    author_comments = models.CharField(max_length=20000, default='')
-    # Quality assurance fields
-    soundly_produced = models.NullBooleanField(null=True)
-    well_described = models.NullBooleanField(null=True)
-    open_format = models.NullBooleanField(null=True)
-    data_machine_readable = models.NullBooleanField(null=True)
-    reusable = models.NullBooleanField(null=True)
-    no_phi = models.NullBooleanField(null=True)
-    pn_suitable = models.NullBooleanField(null=True)
-    # Editor decision. 0 1 2 for reject/revise/accept
-    decision = models.SmallIntegerField(null=True)
-    decision_datetime = models.DateTimeField(null=True)
-    # Comments for the decision
-    editor_comments = models.CharField(max_length=20000)
-    auto_doi = models.BooleanField(default=True)
-
-    def set_quality_assurance_results(self):
-        """
-        Prepare the string fields for the editor's decisions of the
-        quality assurance fields, to be displayed. Does nothing if the
-        decision has not been made.
-        """
-        if not self.decision_datetime:
-            return
-
-        resource_type = self.project.resource_type
-
-        # See also YES_NO_UNDETERMINED in console/forms.py
-        RESPONSE_LABEL = {True: 'Yes', False: 'No', None: 'Undetermined'}
-
-        # Retrieve their labels and results for our resource type
-        quality_assurance_fields = self.__class__.QUALITY_ASSURANCE_FIELDS[resource_type.id]
-
-        # Create the labels dictionary for this resource type
-        labels = {**self.__class__.COMMON_LABELS, **self.__class__.LABELS[resource_type.id]}
-
-        self.quality_assurance_results = []
-        for f in quality_assurance_fields:
-            qa_str = '{} {}'.format(labels[f], RESPONSE_LABEL[getattr(self, f)])
-            self.quality_assurance_results.append(qa_str)
-
-
-class CopyeditLog(models.Model):
-    """
-    Log for an editor copyedit
-    """
-    content_type = models.ForeignKey(ContentType, on_delete=models.CASCADE)
-    object_id = models.PositiveIntegerField()
-    project = GenericForeignKey('content_type', 'object_id')
-    # Either the time the project was accepted and moved into copyedit
-    # from the edit stage, or the time it was reopened for copyedit from
-    # the author approval stage.
-    start_datetime = models.DateTimeField(auto_now_add=True)
-    # Whether the submission was reopened for copyediting
-    is_reedit = models.BooleanField(default=False)
-    made_changes = models.NullBooleanField(null=True)
-    changelog_summary = models.CharField(default='', max_length=10000, blank=True)
-    complete_datetime = models.DateTimeField(null=True)
-
 
 class LegacyProject(models.Model):
     """
